requirements/base.txt
aioredis==1.3.1
amqp==2.6.0
aniso8601==7.0.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.6
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.6.0
django-cors-headers==3.4.0
django-filer==1.7.1
django-graphql-jwt==0.3.1
django-js-asset==1.2.2
django-money==1.1
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-polymorphic==2.1.2
django-querysetsequence==0.13
easy-thumbnails==2.7
future==0.18.2
graphene==2.1.8
graphene-django==2.12.1
graphene-file-upload==1.2.2
graphene-subscriptions==1.0.2
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
kombu==4.6.11
msgpack==1.0.0
phonenumberslite==8.12.7
Pillow==7.2.0
progressbar2==3.51.4
promise==2.3
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMuPDF==1.17.4
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
vine==1.3.0
zipp==3.1.0
zope.interface==5.1.0
